Changed the class, function structure
