+++ gapopulation.py
@@ -1,3 +1,21 @@
+import numpy as np 
+from asyncio import subprocess
+import random
+import math
+from bisect import bisect_right
+import matplotlib.pyplot as plt
+from tkinter import W # right most number we can insert the number
+#import matplotlib.pyplot as plt
+import subprocess
+import sys
+import os
+import fileinput
+from cost_calculator import calculate_bill,calculate_total_cost
+from battery import calculate_cyclelife
+from battery import battery_degradation_cost,calculate_soc_level
+from Input import getInput
+from loads import calculate_shiftable_load_consumption, calculate_shedding_results, calculate_total_load
+
 class Population:
     @staticmethod
     def mutation(individual, cost_function):
